migrating UID methods to core module
